[BUG] fix `get_fitted_params` in case of vectoriztion for forecasters, transformers (#4105)
This PR fixes an unreported bug that was noticed here: https://github.com/sktime/sktime/discussions/4101#discussioncomment-4679066

The bug causes only the diagonal combinations of rows/columns to be present as string keyable values in the fitted params dict in case of vectorized forecasters and transformers.

What should be present is the cartesian product, not the diagonal. This is solved by correctly using `product` instead of `zip` that was used accidentally.

Also fixes an issue (not necessarily bug) with the convention of constructing the strings - they should now always be a string that evaluates to a correct data frame `loc` access command, which previously was not the case for `str` keys (as that would eat away the quotes in `str` coercion).