[BUG] Fix `OptionalPassthrough` `X_inner_mtype` tag (#4115)
Fixes https://github.com/sktime/sktime/issues/4114.

The bug occurs when the `OptionalPassthrough` is applied in a case that requires broadcasting. Through interaction with the `_DelegatedTransformer`, this can lead to wrong conversions.

The fix is delegating all conversions to the wrapped transformer (by removing the override of the `X_inner_mtype` tag and letting all mtypes through).

Adds two tests for the failure case in #4114 (column broadcast) and the parallel failure case for rows.