[ENH] refactor - localize broadcasting in `VectorizedDF` (#4132)
This PR moves functionality for broadcast/vectorized application of estimator methods to `VectorizedDF`, to a new method called `vectorize_est`. It also serves to explore refactor end states.

Reasons to move broadcast/vectorized application to `VectorizedDF`:

* duplication between transformers and forecasters which have very similar logic implemented
* current state can be seen as a violation of the law of Demeter; target state improves cohesion, reduces coupling between base classes and `VectorizedDF`
* localized logic in `VectorizedDF` seems to be a more natural starting point for adding `dask` and scaling/distributed features, see https://github.com/sktime/sktime/issues/4013

Also contains a direct unit test for `vectorize_est`.
Note that the new method `vectorize_est` is also indirectly covered by tests for vectorization functionality in transformers and forecasters, which also covers the external points of the refactor in `BaseForecaster` and `BaseTransformer`.