[ENH] fixes for forecasters to retain `name` attribute in `predict` (#4161)
This PR collects fixes for a number of forecasters to preserve the `name` attribute in `predict`.
See #4144. The issues were found by #4157.

Does *not* contain changes to the conversion logic, or tests (e.g., #4157).
Testing is via merging this PR into the options dealing with the `name` attribute.

Contains fixes for:

* `ARDL`
* `ARIMA` and `AutoARIMA`
* `AutoEnsembleForecaster`
* `AutoETS`
* `Croston`
* `NaiveForecaster`
* `OnlineEnsembleForecaster`
* reducers
* `SquaringResiduals`
* `StackingForecaster`
* `StatsForecastAutoARIMA`
* `STLForecaster`
* `ThetaModularForecaster`
* `TrendForecaster` and `PolynomialTrendForecaster`

As tempting as it might be to add the fix to `BaseForecaster` somewhere in `predict`: whilt it might be DRY-er, I think it would muddle concerns. If at all, this should go in `convert`-like boilerplate rather than in the base class itself, but unclear how that would look like.