[BUG] fix race condition in `tsfile` tests (#4192)
Split out of https://github.com/sktime/sktime/pull/4140

This fixes a race condition in tests, if run in parallel. The "./Temp" dir might get deleted while another test is relying on its presence. This fix replaces the explicit creation of a temp dir with a fixture, unique to each test execution.