[MNT] Remove restrictions on branch for workflow that autodetect and updates `CONTRIBUTORS.md` (#4323)
Reference PR #4257 and issue #4103

We're simply removing the restriction on branches which would enable the workflow to be triggered whenever changes are detected to `.all-contributorsrc` regardless of the branch name. Please see the discussion in #4257 and #4103 for the complete context.

After we merge this PR, I will be able to resume testing the workflow via #4257 and we can decide if we want to keep this workflow since it only reduces manual work with no apparent cons.