[DOC] updated `dtw_distance` docstring example to include import (#4324)
Adds to #4264

This fixes a small pet peeve of mine with the current documentation, which is that the actual import for the technique is often not listed in the example, so you can't just copy and paste from the API reference to see how it works.  This inconvenience is compounded by the fact that the actual imports you need to use are often irregular and can't be easily inferred from the way different methods are listed in the API docs.  So you often have to go on a bit of a chase just to figure out how to import something.