[ENH] conditional transform (#4248)
See https://github.com/sktime/sktime/issues/4249

This adds a transformer compositor which executes a transform conditionally, e.g., deseasonalizes if seasonality is detected, or differences if stationarity is detected. The condition can be specified flexibly based on any estimator (with a `fit` interface), and the if/else transformation can also be configured.

Relies on https://github.com/sktime/sktime/pull/4360 for the seasonality example.