[ENH] turn `BKFilter` into a direct interface (#4346)
Following discussion in https://github.com/sktime/sktime/pull/4342, I noticed that the `BKFilter` did not interface, but in fact copy the code from `statmodels`.

This was an oversight in review, apologies. The code should interface (not copy) the statsmodels function, to avoid duplication and forking.

The fix should not change any functionality.

Also adds a second set of test parameters.