[BUG] fix failure in `VECM.predict_interval` if multiple coverage values were passed (#4411)
See also #4394 (relevant comment: https://github.com/sktime/sktime/pull/4394#issuecomment-1489052793).

The previous implementation of `VECM` caused `ValueError` when a list of `coverage` values (length 2 or more) was passed to `predict_interval`. This was caused because lower/upper bounds determined for different coverages were not stored, and it was trying to match the last iteration with all columns generated independently. This PR addresses that by combining the results from all the different coverage values, and then creating the final results.