[ENH] remove superfluous implementation checks in `_predict_interval` and `_predict_quantiles` of `BaseForecaster` (#4481)
In the existing implementation of `_predict_interval` in `BaseForecaster`, first it checks whether any of `_predict_interval`, `_predict_quantiles` or `_predict_proba` is overridden or not. But the check on `_predict_interval` would never evaluate to `True` when called from an inherited forecaster, as if it is indeed overridden, that overridden method would have been called, and not that of the `BaseForecaster`.

Similar arguments of can be made for `_predict_quantiles` in `BaseForecaster` as well, where using `_predict_interval` is sufficient.

This PR attempts to solve the above arguments by removing these unnecessary checks.