[BUG] fix input check error message in `BasePairwiseTransformerPanel` (#4499)
This PR fixes an unreported bug where the informative error message from
a failed input check in `BasePairwiseTransformerPanel` would not be
displayed, instead a cryptic exception about `NoneType` would be raised.

The failure is due to a return of `check_is_mtype` being `None` in the
case where the check fails, but being treated as a `dict` erroneously
(which it is in the success case). This was fixed.