&achieveordie [BUG] fix bug from clash between `ABC` inheritance and RNN `fit` override (#4527)
By accident, we merged https://github.com/sktime/sktime/pull/4185
without re-running the tests - this masked a bug. Which is not too grave
as it is post-release, and the release CI has all passed.

Details: on the new `main`, the merge caused a clash with the `ABC`
inheritance in `BaseRegressor` as one of the methods was not overridden.
Also, `fit` should be `_fit`, which is also fixed.