[ENH] Extend forecasting benchmarking framework to multiple metrics, add test coverage (#4586)
I recently looked through the changes made to the benchmarking module in
sktime because I'm interested in extending sktime benchmarking with
`kotsu`. As I went through the codebase, I noticed a comment below in
the code, that there was a need to rewrite `evaluate` to handle multiple metrics
but `evaluate` indeed already supports this. we just need to give the
scoring argument as a list of `BaseMetric` and it will compute multiple
metrics without having to refit a model n number of times for n metrics.

I've also added an extra test coverage to ensure this works properly.