[BUG] fix incorrect sorting of `n_best_forecasters_` in `BaseGridCV` if metric's `lower_is_better` is `False` (#4590)
Fixes the following bug in `BaseGridCV`.

On line 221, a rank column is populated based on
scoring tag `lower_is_better`. A few lines of code later, BaseGridCV
updates its n_best_scores and n_best_forecasters. so to do this, the
same table is then sorted based on the rank column created in line 221,
but with ascending order following the scoring tag again. This works
well for metrics where lower is better, which is the case for all
forecasting metrics as far as I know lol (minimum is better). However,
it would be incorrect for the opposite case. To give you an example:

| column_1  | rank |
| ----------- | ---- |
| 0.527870   | 2.0  |
| 0.677122   | 1.0  |
|  0.060541  | 3.0  |

if ascending = False, the tag for when higher is better, the algorithm
is fitting the n **worst** forecaster rather.
| column_1  | rank |
| ----------- | ---- |
|  0.060541  | 3.0  |
| 0.527870   | 2.0  |
| 0.677122   | 1.0  |