[BUG] fix for `get_fitted_params` in `_HeterogenousMetaEstimator` (#4633)
This fixes a bug in `get_fitted_params` of `_HeterogenousMetaEstimator`,
which is at the root of https://github.com/sktime/sktime/issues/4574.
Also fixes #4574.

`get_fitted_params` of `_HeterogenousMetaEstimator` was accidentally
calling the private interface point of components rather than the public
interface point that it should have.

For reviewers: the first call should be to `_get_fitted_params` of the
base class, which is correct. Subsequent calls to components should be
to the public interface, `get_fitted_params`.

Contains a regression test for #4574.