[ENH] set `ForecastX` missing data handling tag to `True` to properly cope with future unknown variables (#4876)
This PR sets the `ForecastX` missing data handling tag to `True` to
properly cope with future unknown variables.

An edge case - or, arguably, a general case if none of the future
variables are known - is showcased in this issue discussion:
https://github.com/sktime/sktime/issues/4848

More precisely, `ForecastX` should always allow for `nans` in future
`X`, as these may be in places where there is genuinely no knowledge
about the future. Alternatively, dummy values have to be provided which
sounds odd.