[ENH] add Monash Forecasting Repository data loader (#4826)
See #4695. Since this PR implement one of the two repos mentioned in the
reference PR, it should'nt be closed.

Add a `fetch_forecasting` dataloader that retrieves forecasting datasets
from the Monash Time Series Forecasting Repository. This dataloader
first looks for the dataset in the extract_path. If the dataset is not
present, it attempts to download the data from
*https://forecastingdata.org/* and saves it to the extract_path.

To ensure consistency with the implementation of `load_UCR_UEA_dataset`,
a dataloader that does the same for classification datasets,
`load_forecastingdata` closely follows it and uses some of the private
functions that it relies on.

Discussed: A naming fetch instead of load would be better to distinguish
it from other dataloaders that load data from sktime pre-installed
datasets as per suggestion in #4314. For now, we've called it
`load_forecastingdata` for consistency reasons.