[ENH] replace `"Coverage"` and `"Quantiles"` default variable name in univariate case with variable name (#4880)
Copy of https://github.com/sktime/sktime/pull/4780 which was reverted to
ensure it ships with 0.21.0, not 0.20.1.

Fixes #4709 - in `predict_interval` and `predict_quantiles` return,
replaces `"Coverage"` and `"Quantiles"` default variable name in the
univariate case with the variable name, to make this consistent with the
multivariate case. If no variable name is available, the variable name
used is the integer 0.

This is done with a deprecation horizon or two MINOR cyles, by
introducing a `legacy_interface` argument to public and private
`predict_interval`, `predict_quantiles`. A value `True` (default) gives
current behaviour, `False` gives future behaviour. In 0.22.0, we switch
the default to `False`, and remove the argument in 0.23.0.

This includes tests for `legacy_interface` both `True` and `False`, with
deprecation notes on removing this later.