[ENH] `HolidayFeatures` transformer (#4909)
Fixes #1675 

This PR adds a new `HolidayFeatures` transformer, which extracts holiday features from datetime
index of the input dataframe. It expects a dictionary
of [date, string] pairs (which could also be a HolidayBase object from
holidays package) and returns a dataframe with the
features in the required format (which could be either be dummies,
categorical, indicator or any combination of three).
This PR includes the preliminary work done in #1722

#### Does your contribution introduce a new dependency? If yes, which
one?

It has a soft dependency on "holidays" package.

#### What should a reviewer concentrate their feedback on?

A reviewer should concentrate their feedback on:
- The existing features of the transformer.
- Any feature to add or remove.
- Any other suggestions

#### Did you add any tests for the change?

Yes, tests are added as test_holidays.py in the respective section.