&MCRE-BE [BUG] Fix `Differencer.inverse_transform` on train data if `na_handling=fill_zero` (#4998)
In `Differencer`, set first lag values to nan if it has the same index as the train time
series. This should fix the strange results in `inverse_transform` of the
Differencer if fill_zero is set.

Fixes #4993 and fixes #4965.

The fix sets the first lag values of the input time series to nan if the
index aligns with the index of the train time series. Thus, the
`combine_first` operation will set it to the values of the original time
series, and `cumsum` reproduces the original time series.

Also added a test for reproducing the original bug.