[ENH] Add DropNA transformer (#5049)
#### Reference Issues/PRs
Fixes #4977.

#### What does this implement/fix? Explain your changes.
Implements a `DropNA` transformer that saves the dropped index/column
names from `fit`.

#### What should a reviewer concentrate their feedback on?
Any feedback is appreciated. This is mostly a thin wrapper around
`pd.DataFrame.dropna`, accepting the arguments `axis`, `how`, and
`thresh`, defaulting to `axis=0` and `how=any`. With univariate series,
only `axis=0` will work.

I am validating the inputs to some degree and I've extended the `thresh`
argument to also accept a fraction of non-NA values (pandas only accepts
counts). I've always found the pandas default of specifying the
threshold in terms of non-NA values strange (as opposed to specifying NA
values) , but I've left it as is to be consistent. Thoughts on this?

#### Did you add any tests for the change?
I've added a couple of tests for the functionality and to make sure the
parameter validation works. I haven't added any tests with a univariate
series so far.