[BUG] fix `ConformalIntervals` failure if wrapped estimator supports hierarchical mtypes - restriction to Series (#5091)
This PR fixes #5092, i.e., the `ConformalIntervals` failure where the
wrapped estimator supports hierarchical or panel mtypes.

As discussed in the issue, the failure condition was if the wrapped
estimator supported a hierarchical mtype. As the tags were cloned and
`ConformalIntervals` did not support hierarchical or panel mtypes, this
led to failure.

This is fixed by no longer cloning the mtype tags, but letting all
mtypes through.

This PR also cleans up the mtype lists exported from the `registry`,
which - despite the name, weren't lists but `numpy.ndarray`-s. This
would also cause problems if they were directly used in tags. This has
been rectified, the objects are now lists (as the name claims and as the
user would expect).