[BUG] fix `ConformalIntervals` failure if wrapped estimator supports hierarchical mtypes - genuinely hierarchical (#5093)
This PR fixes #5092, i.e., the `ConformalIntervals` failure where the
wrapped estimator supports hierarchical or panel mtypes.

As discussed in the issue, the failure condition was if the wrapped
estimator supported a hierarchical mtype. As the tags were cloned and
`ConformalIntervals` did not support hierarchical or panel mtypes, this
led to failure.

This is fixed by adding vectorization functionality for panel and
hierarchical scitypes.

This lets genuinely hierarchical predictions through, but does the
conformal interval computation per individual instance.

Depends on https://github.com/sktime/sktime/pull/5091, which contains
the test for #5092, and a clean-up of the `datatypes` registry.