[BUG] fix `numba` based aligners (#5116)
This PR fixes a bug in the `numba` based aligners which were detected by
the release action.

The bug was unreported, and causes `transform` to break at call, due to
the `X_inner_mtype` tag and the handling of `X` in `transform` being
inconsistent. This has now been rectified, assuming `numpy3D` in both
tag and `_transform` logic.