[MNT] 0.22.0 release action - deprecate `deprecated` in 0.21.0, remove in 0.22.0 (#4822)
In the spirit of minimizing dependencies, I suggeste to deprecate
(0.21.0 or 0.20.1) and then remove the `deprecated` package as a
dependency.

We haven't used `deprecated` for deprecations in a while, especially
since interfaces have matured and we rarely deprecate entire classes or
functions anymore.

Deprecations are still common, but they mostly affect arguments,
parameters, or their behaviour instead, cases not covered by the
`deprecated` package afaik.

Of course, as a current core dependency, we would need to deprecate
`deprecated` itself, as a dependency. Following the established
procedure, this would be announced as a deprecation in the next release
(0.20.1 and/or 0.21.0), and then removed in 0.22.0.

Depends on https://github.com/sktime/sktime/pull/4824 which removes a
forgotten import of `deprecated`.