[MNT] move fixtures in `test_dropna` to `pytest` fixtures (#5153)
This PR moves fixtures in `test_dropna` to `pytest` fixtures.

Thanks again, @hliebert, for the great transformer!

I noticed a few things that could be bugs in the tests, would appreciate
your quick eye on this:

* in current line 64, there is a `dropna` that does nothing, as the
default is not mutating (`inplace=False`). Did you want an assignment?
* in `test_dropna_conflicting_arguments`, the `remember` arg was not
used. Should the transformer be constructed with it?