[ENH] Allow unrestriced ID string for BaseBenchmarking (#5130)
Towards #4873 

The current state of kotsu benchmarking is that it restrict users to
only input ID of certain format i.e ID that follows the following format
`[username/](entity-name)-v(major).(minor)`

We should be flexible and allow users to decide if they want to use this
feature rather than enforcing it upon them. Therefore, this PR aims to:
1. Set the default option to switch off the input ID format.
2. Allow users to define their own ID format using regex if they want
to.

This is also a design for which to override, make patches and get more
control to kotsu framework within sktime. I started with a small change
like this one first to get a green light before proceeding with others.

1. Some of the existing tests are changed to reflect the default ID
format that no longer contains "-v1" at the end.
2. Added a new test to ensure an error message is raised if a user
specifies ID format.