[ENH] fix `test_set_freq_hier` for `pandas 2.1.0` (#5185)
`test_set_freq_hier` fails since `pandas 2.1.0` as the condition for the
test - a `freq` is set to `none` - is not anymore established. This is
due to a change in `pandas` that preserves this attribute more widely.

To ensure equivalent test coverage, the fix is to set the `freq` to
`none` manually in the relevant case.