[BUG] fix BaseWindowSplitter get_n_split method for hierarchical data (#5012)
Fixes #4971

The logic of the `BaseWindowSplitter.get_n_split` method does not work
for hierarchical cases. This is mainly because it counts cutoffs on y,
assuming that it is either a **single** or a **multivariate** time
series. Therefore, when hierarchical data is passed, the logic breaks
because it counts cutoff as if the "pd_multiindex_hier" is a single
series when it is not.

Solution
Using the existence implementation, for when pd_multiindex_hier datatype
is passed, the cutoff point is counted based on the first instance of
the lowest level time series hierarchy.

Other comments
I also found out that other splitters, like `CutoffSplitter`, may return
the wrong number of splits because they use the same logic as
`BaseWindowSplitter` to count the splits. Also, It seems that the
original implementation for hierarchical splitter #2599 may not consider
how `get_n_splits` method should return in the case of hierarchical.