[ENH] make `statsforecast` adapter compatible with optional `predict` `level` arguments, and different init param sets (#5112)
Fixes #5024 and #5207:

* Versions previous to v1.5.0 of `statsforecast` do not have the `level`
argument in the `predict` or `predict_in_sample` methods of the
forecasters.
* `statsforecast` removed the `parallel` argument in some estimator since v1.6.0

Details:

Now before calling `predict` or `predict_in_sample` we make sure the
methods contain the `level` argument otherwise we do not pass it.

To handle the `level` argument we will set tags dynamically at
instantiation time to see if they are there in the methods
`predict_in_sample` and `predict`.

Also newer versions of `statsforecast` removed parameters that were
available in previous versions. We will take care of it by throwing a
warning if the user passes a keyword argument that is not supported by
the current version of `statsforecast` that they might be using. We will
throw a warning if the user passes the keyword argument which is
non-existent in its current version of `statsforecast`.

If a parameter is set by default and NOT passed by the user, the
parameter will also be ignored but NO warning will be thrown as the user
did not intend to use that parameter in the first place, so he should
not care.