[ENH] ensure `ForecastingPipeline` is compatible with "featurizers" (#5252)
If there is are exogenous variables, `ForecastingPipeline` would fail
with featurizers that add exogenous variables, e.g., `YtoX`, or
`FourierFeatures`, or `DatetimeFeatures`.

This PR solves that issue, by ensuring that `ForecastingPipeline` passes
a zero-column `pd.DataFrame` with the right `pd.Index`, as `y`, to the
featurizers. Previously, `y=None` would be passed, preventing the
featurizers' access to the `fh` index.

Also adds a test.