[ENH] consolidating splitters as their own module with systematic tests and extension (#5017)
Towards #4889 

- [x] Create a new module, `sktime.split`, which will contain all the
splitters
- [x] Create submodule `base` which only has the `BaseSplitter` and
`BaseWindowSplitter`
- [x] Move the splitters into individual files inside `split`.
`__init__` should export them all. The old location should be subject to
deprecation - imports should still work but raise a warning.
- [x] Create a test module, `split.tests`. All current tests should move
there.
- [ ] Add an extension template for splitters, in `extension_templates`
- this will inform tests below as it will make concrete the key methods.
- [ ] Create a new test class, `tests::TestAllSplitters`. This should
inherit from `TestAllObjects`, and be similar to `TestAllAligners`,
`TestAllDistributions`, and similar. It should also be registered in
`check_estimators` (see there for pattern)
- [ ] The generic tests should be migrated to `TestAllObjects`, as long
as it is a test for all splitters.
- [ ] Think about new tests, e.g., "test that `get_n_splits` is always
the same as length of `split`". All methods of splitters should be
covered.
- [ ] It may make sense to add test scenarios, although this may equally
be covered by simple fixtures.