[BUG] Allow `evaluate` to accept any combination of multiple metrics with correct predict method (#5192)
Fixes #5167 and #5194


The current `evaluate` logic only runs predict method based on the first
metric in the list. The logic will break if users supplement `evaluate`
with a list of different metrics type, i.e metrics used to evaluate
different predict methods.

This PR solves it by:
1. Ensuring that an estimator is fitted once
2. Once in the fitted state, use the correct `predict` method that
corresponds to the metrics supplied.

This PR also improves the estimation of prediction time by tracking the
time only when an **estimator is making predictions**, I excluded the
setup time, i.e lookup time for finding the correct `predict` method and
extracting metrics tag and metrics arg that were previously included in
the prediction time measurement which I think should not be included -
although this may only take a few ms.