[BUG] in `Imputer`, fix `y` not being passed in `method="forecaster"` (#5287)
This fixes #5284.

The reason that `y` was not passed is the `y_inner_mtype` tag which
always said that internally `y` is not needed (or passed). However,
`method="forecaster"` may require it as example #5284 shows, and this is
a bug. The fix is setting the tag correctly, in this case.

Adds the failure case from #5284 as a test.