[ENH] in scitype check, replace base class register logic with type tag inspection (#5288)
This replaces the scitype check based on base class inheritance with a
simple tag.

Why is this needed: we are moving to an interconnected ecosystem of
packages, and the same interface may be present in multiple. Checking by
inheritance will check against the class in a specific package though,
making the check not extensible to multiple packages - an example is the
issue with distributions being present in both `sktime` and `skpro`, see
https://github.com/sktime/sktime/pull/5271

This is fixed by adopting the tag based approach which `sklearn` is now
also using:

* each base class has an `object_type` tag which has the string or
strings which was previously in the base class register
* the `scitype` type checking function now inspects that tag instead of
checking inheritance.

This also has the advantage of removing coupling to the base class
register.

For now, the change does not need deprecation, as the tags follow the
same inheritance system as classes do (from `scikit-base`), therefore
downstream users or extenders inheriting from `sktime` classes will not
be affected.

The `scitype` function is kept deprecation safe by running the legacy
logic if the new tag cannot be found.