[BUG] fix `temporal_train_test_split` for hierarchical and panel data in case where `fh` is not passed (#5330)
This PR fixes #5107 by replacing the inner workings of
`temporal_train_test_split` with a splitter, making the design both DRY
and reliant on splitter boilerplate.

This is done as follows:

* addition of a new splitter `TemporalTrainTestSplitter`, which
encapsulates the `temporal_train_test_split` logic
* internally, `TemporalTrainTestSplitter` is a re-implementation of the
`temporal_train_test_split` logic as a splitter class, giving access to
the `BaseSplitter` boilerplate.
* removed osuleaf example, as that was time series classification and
got split across instances in the sample rather than time points. This
would be inconsistent with the general splitter interface and the
docstring specification.

The issue with the panel data is solved by recurring to the existing
broadcasting boilerplate of `BaseSplitter`.

Depends on https://github.com/sktime/sktime/pull/5332 for testing.