[DOC] Correct code block formatting for pre-commit install command (#5377)
#### Reference Issues/PRs
Fixes #5308

#### What does this implement/fix? 
This PR corrects the code block formatting for the `pre-commit install`
command in the developer guide. The command was not properly formatted
as a code block, which could cause confusion for readers. The command is
now enclosed in a code block, making it clear that it's a command to be
executed in the terminal.