[ENH] `ForecastX` enable fitting `forecaster_y` on forecasted `X` (#5334)
Enable `ForecastX` to train `forecaster_y` on prediction of `forecaster_X` to
enable `forecaster_y` to consider the uncertainty of the forecasted `X`.

#### Reference Issues/PRs
fixes partially #5225

#### What does this implement/fix? Explain your changes.
This PR adds a parameter to `ForecastX` that enables to control if
ForecastX should fit the internal `forecaster_y` on the ground truth data
of `X` or of the forecasted `X`. Fitting on the forecasted `X` enables
`forecaster_y` to account the uncertainty associated with forecast.


#### Did you add any tests for the change?
Yes.