[MNT] Migrate from `pykalman` to `pykalman-bardo` (#5277)
Consider migration from `pykalman` to `pykalman-bardo`, as
[pykalman](https://github.com/pykalman/pykalman) project is no longer
maintained. There were some issues that were fixed, see:
https://github.com/pybardo/pykalman/blob/v0.9.6/CHANGELOG

I'm going to maintain
[pykalman-bardo](https://github.com/pybardo/pykalman), react to issues,
and fix bugs. For now the API is the same as in the initial package, but
it might evolve in time.

I hope you will find it useful for your project!

#### Reference Issues/PRs
`pykalman` compatibility with python 3.11

#### Does your contribution introduce a new dependency? If yes, which
one?
`pykalman`-> `pykalman-bardo`